+++ login-workflow/src/contexts/index.tsx
@@ -3,3 +3,4 @@ export * from './RegistrationContext';
 export * from './SharedDictionaries';
 export * from './ErrorContext';
 export * from './AuthContext';
+export * from './OktaAuthContext';